+++ backend/Loan/Loan.DAL/Entities/Loan.cs
@@ -4,20 +4,21 @@ using Common.Persistence;
 
 namespace Loan.DAL.Entities;
 
-public class Loan: BaseEntity {
+public class Loan : BaseEntity
+{
     [Key]
     public Guid Id { get; set; } = Guid.NewGuid();
-    
+
     public Guid UserId { get; set; }
-    
+
     public Guid AccountId { get; set; }
-    
+
     public Tariff Tariff { get; set; }
     public List<PaymentRequest> Payments { get; set; } = [];
     public DateTime? LastChargeDate { get; set; }
-    
+
     public CurrencyType CurrencyType { get; set; }
     public bool IsClosed { get; set; } = false;
-    
-    public int Debt { get; set; }
+
+    public long Debt { get; set; }
 }
