+++ backend/OperationHistory/OperationHistory.DAL/Migrations/OpHistoryDbContextModelSnapshot.cs
@@ -31,8 +31,8 @@ namespace OperationHistory.DAL.Migrations
                     b.Property<Guid>("AccountId")
                         .HasColumnType("uuid");
 
-                    b.Property<int>("Amount")
-                        .HasColumnType("integer");
+                    b.Property<long>("Amount")
+                        .HasColumnType("bigint");
 
                     b.Property<DateTime>("CreatedAt")
                         .HasColumnType("timestamp with time zone");
@@ -82,8 +82,8 @@ namespace OperationHistory.DAL.Migrations
                     b.Property<Guid>("AccountId")
                         .HasColumnType("uuid");
 
-                    b.Property<int>("CalculatedAmount")
-                        .HasColumnType("integer");
+                    b.Property<long>("CalculatedAmount")
+                        .HasColumnType("bigint");
 
                     b.Property<DateTime>("CreatedAt")
                         .HasColumnType("timestamp with time zone");
